prompt-vault/backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.3

# Supabase
supabase>=2.0.0

# Firestore
google-cloud-firestore>=2.13.0

# Gemini API
google-generativeai>=0.3.0

# HTTP client
httpx>=0.25.1
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.1

